+++ example/src/containers/App/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import StateTemplate from 'state-template';
+import StateTemplate, { history } from 'state-template';
 
 import routes from './routes';
 
-const App = () => <StateTemplate routes={routes} />;
+const App = () => <StateTemplate routes={routes} history={history} />;
 
 export default App;