+++ example/src/containers/App/routes.js
@@ -1,4 +1,5 @@
 import Home from '../Home';
+import OtherRoute from '../OtherRoute';
 
 // order here determines order in navigation menu
 export const routesByKey = {
@@ -9,6 +10,14 @@ export const routesByKey = {
     exact: true,
     icon: 'ca-gov-icon-home',
     component: Home,
+  },
+  otherRoute: {
+    key: 'other',
+    name: 'Other Route',
+    path: '/other',
+    exact: true,
+    icon: 'ca-gov-icon-home',
+    component: OtherRoute,
   },
 };